requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
auditwheel==5.1.2
backcall==0.2.0
beautifulsoup4==4.11.1
brotlipy==0.7.0
cachetools==5.3.0
carla==0.9.10
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
comm==0.1.2
ConfigArgParse==1.5.3
configparser==5.3.0
cryptography==38.0.4
cycler==0.11.0
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
dictor==0.1.11
distro==1.7.0
docker-pycreds==0.4.0
elementpath==1.3.3
ephem==4.1.3
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.6.0
Flask==2.2.2
flit_core==3.6.0
fonttools==4.33.2
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gdown==4.4.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
gql==0.2.0
graphql-core==1.1
grpcio==1.51.1
gym==0.21.0
gym-notices==0.0.8
idna==3.4
imageio==2.25.0
imgaug==0.4.0
importlib-metadata==4.13.0
importlib-resources==5.10.2
ipykernel==6.21.1
ipython==8.9.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.2
lightning-utilities==0.6.0.post0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.4
nbformat==5.5.0
nest-asyncio==1.5.6
networkx==2.2
numpy==1.24.2
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
open3d==0.16.0
opencv-python==4.2.0.32
packaging==21.3
pandas==1.5.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
plotly==5.13.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-trees==0.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pyelftools==0.28
pygame==2.1.2
Pygments==2.14.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyquaternion==0.9.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==0.7.1
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
sentry-sdk==1.15.0
setproctitle==1.3.2
setuptools==65.6.3
Shapely==1.6.4.post2
shortuuid==1.0.11
simple-watchdog-timer==0.1.1
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.6.2
subprocess32==3.5.4
tabulate==0.8.9
tenacity==8.2.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2023.2.3
timm==0.4.5
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.11.1
torchvision==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.9.0
typing_extensions==4.2.0
urllib3==1.26.14
wandb==0.8.29
watchdog==2.2.1
wcwidth==0.2.6
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==4.0.5
xmlschema==1.0.18
yarl==1.8.2
zipp==3.13.0
